Add files via upload
Title of the project: NYC Subway Stops
Description of the project: Tracks stops of all NYC subway stops. 
How to Run: Press the button in the top left to start the animation. 
Roadmap of future improvements: Creating a lined path between stops so it is visually easy to see where the subway will be stopping at in the future and in the past. 
License information: This was inspired and partially curated with the help of MIT xPro